requirements.txt
streamlit
pandas
scikit-learn
sentence-transformers
Pillow
requests

